The code now (theoretically) should run and generate the video as expected. Haven't tested it yet

+++ run.sh
@@ -10,6 +10,8 @@ CONFIG_FILE="$SCRIPT_DIR/config/config.ini"
 OUTPUT_FOLDER=$(awk -F " = " '/OutputFinalFolder/ {print $2}' "$CONFIG_FILE")
 OUTPUT_TEX_FILE=$(awk -F " = " '/OutputTeXFile/ {print $2}' "$CONFIG_FILE")
 OUTPUT_SPEECH_FILE=$(awk -F " = " '/OutputSpeechFile/ {print $2}' "$CONFIG_FILE")
+SEPARATE_PAPERS_FOLDER=$(awk -F " = " '/SeparatePapersFolder/ {print $2}' "$CONFIG_FILE")
+OUTPUT_VIDEO_FILE=$(awk -F " = " '/OutputVideoFile/ {print $2}' "$CONFIG_FILE")
 
 # Ensure OUTPUT_FOLDER is evaluated for variables such as $HOME
 eval OUTPUT_FOLDER=$OUTPUT_FOLDER
@@ -17,13 +19,20 @@ eval OUTPUT_FOLDER=$OUTPUT_FOLDER
 # cd into the script directory, if haven't done that already
 cd "$SCRIPT_DIR"
 
+# Clean up the previous tex, mp3 and mp4 files
+rm -f "./workdir/$SEPARATE_PAPERS_FOLDER/*"
+
 # Run the script
 source ./env/bin/activate
 python ./src/main.py
 
+# This script will create the video for YouTube
+./src/compile_all_papers.sh "$SEPARATE_PAPERS_FOLDER" "$OUTPUT_VIDEO_FILE"
+
 # Move the files to the output folder
-mv "./workdir/$OUTPUT_TEX_FILE" "$OUTPUT_FOLDER"
-mv "./workdir/$OUTPUT_SPEECH_FILE" "$OUTPUT_FOLDER"
+cp "./workdir/$OUTPUT_TEX_FILE" "$OUTPUT_FOLDER"
+cp "./workdir/$OUTPUT_SPEECH_FILE" "$OUTPUT_FOLDER"
+cp "./workdir/$SEPARATE_PAPERS_FOLDER/$OUTPUT_VIDEO_FILE" "$OUTPUT_FOLDER"
 
 # Compile the document
 cd "$OUTPUT_FOLDER" || { echo "Output folder doesn't exist! Exiting..."; exit 1; }
@@ -31,7 +40,7 @@ pdflatex -interaction=nonstopmode "$OUTPUT_TEX_FILE"
 
 # Cleanup the folder - only leave .tex, .pdf, and .mp3 files
 for file in *; do
-    if [[ ! $file =~ \.tex$ ]] && [[ ! $file =~ \.pdf$ ]] && [[ ! $file =~ \.mp3$ ]]; then
+    if [[ ! $file =~ \.tex$ ]] && [[ ! $file =~ \.pdf$ ]] && [[ ! $file =~ \.mp3$ ]] && [[ ! $file =~ \.mp4$ ]]; then
         rm "$file"
     fi
 done
